requirements_win.txt
concurrent_log_handler==0.9.25
dash==2.18.2
feffery_antd_components==0.3.15
feffery_utils_components==0.2.0rc27
feffery_antd_charts==0.1.5
pycryptodome==3.21.0
PyJWT==2.9.0
user_agents==2.2.0
yarl==1.16.0
feffery_dash_utils==0.1.5
pillow==10.4.0
py-cpuinfo==9.0.0
cacheout==0.16.0
peewee==3.17.8
otpauth==2.1.1
Flask-Compress==1.15

psutil==6.0.0
PyMySQL==1.1.1
rpyc==6.0.1
apscheduler==3.10.4
paramiko==3.5.1
sqlalchemy==2.0.31
waitress==3.0.2